requirements.txt
# --- Dependências de CORE do Sistema (Manter) ---
Django==5.2.7       # Framework Web
         # Manter a versão do Framework Principal
groq            
psycopg2-binary 
redis           

# --- Servidor Web (ASGI/WSGI) ---
gunicorn       
uvicorn[standard]

# --- Observabilidade ---
prometheus_client 

# --- Dependências Comuns e de Rede (LIMPAS) ---
attrs
Babel
blinker
Brotli
certifi
chardet
click
colorama
configobj
cryptography
httplib2
idna
Jinja2
jsonpatch
jsonpointer
jsonschema
markdown-it-py
MarkupSafe
mdurl
mutagen
netaddr
oauthlib
olefile
pillow
pycryptodomex
Pygments
PyJWT
pyparsing
pyrsistent
pyserial
python-dateutil
pytz
PyYAML
requests
setuptools
six
typing_extensions
urllib3
wheel

google-auth
google-api-python-client